requirements.txt
django==2.0.6
django-cors-headers==2.3.0
django-rest-framework==0.1.0
